add configurable rate limiter for jar collector processing

+++ newrelic-agent/src/test/java/com/newrelic/agent/config/JarCollectorConfigImplTest.java
@@ -58,7 +58,7 @@ public class JarCollectorConfigImplTest {
         jarCollectorConfig = new JarCollectorConfigImpl(configProps);
 
         assertEquals(DEFAULT_ENABLED, jarCollectorConfig.isEnabled());
-        assertEquals(DEFAULT_MAX_CLASS_LOADERS.intValue(), jarCollectorConfig.getMaxClassLoaders());
+        assertEquals(DEFAULT_MAX_CLASS_LOADERS, jarCollectorConfig.getMaxClassLoaders());
     }
 
     @Test
